whatsapp_cloud/requirements.txt
# WhatsApp Cloud API – Flask webhook + Groq
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
